requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# UI
streamlit>=1.40.0
requests>=2.32.0

# Async support
httpx>=0.27.0
aiohttp>=3.10.0

# Optional: for additional tools
wikipedia>=1.4.0
duckduckgo-search>=6.3.0
